requirements/runtime.txt
albumentations==1.3.0
einops==0.3.0
huggingface_hub==0.13.4
imageio==2.9.0
mmdet==3.2.0
mmpose==1.2.0
modelscope==1.9.3
omegaconf==2.3.0
opencv_python==4.1.2.30
opencv_python_headless==4.7.0.72
Pillow==10.1.0
pytorch_lightning==1.9.0
PyYAML==6.0.1
safetensors==0.4.0
streamlit==1.20.0
timm==0.9.8
transformers==4.27.4
typing_extensions==4.8.0
scipy==1.9.1
streamlit==1.20.0
timm==0.9.8
tqdm==4.65.0
transformers==4.27.4
open_clip_torch==2.7.0